+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Pass props to the extensionPoint of `SearchPage`.
+- Modify `store/search` path from `/:term/` to `/:term/s`
 
 ## [0.2.5] - 2018-6-14